# [huf] Improve fast C & ASM performance on small data